サービス一覧をチャンネル番号順に変更(※要注意※)
サービスのID算出方法を変更したので自動予約登録と番組表設定の互換性が失われています。
自動予約と番組表は設定しなおす必要があります。

+++ EpgTimer/EpgTimer/SearchWindow.xaml.cs
@@ -741,7 +741,7 @@ namespace EpgTimer
                 {
                     if (info.ServiceName.Equals(item.ServiceName))
                     {
-                        defKey.serviceList.Add((long)info.ID);
+                        defKey.serviceList.Add(info.ID);
                     }
                 }
                 searchKeyView.SetSearchKey(defKey);
